Runs on Linux

+++ ex_argv.h
@@ -11,17 +11,17 @@ char	**argv;
 char	**argv0;
 char	*args;
 char	*args0;
-short	argc;
-short	argc0;
-short	morargc;		/* Used with "More files to edit..." */
+int	argc;
+int	argc0;
+int	morargc;		/* Used with "More files to edit..." */
 
 int	firstln;		/* From +lineno */
 char	*firstpat;		/* From +/pat	*/
 
 /* Yech... */
 struct	glob {
-	short	argc;			/* Index of current file in argv */
-	short	argc0;			/* Number of arguments in argv */
+	int	argc;			/* Index of current file in argv */
+	int	argc0;			/* Number of arguments in argv */
 	char	*argv[NARGS + 1];	/* WHAT A WASTE! */
 	char	argspac[NCARGS + sizeof (int)];
 } frob;
